+++ components/director/internal/open_resource_discovery/processors/common.go
@@ -1,0 +1,21 @@
+package processors
+
+import (
+	"context"
+	"github.com/kyma-incubator/compass/components/director/pkg/log"
+)
+
+func addFieldToLogger(ctx context.Context, fieldName, fieldValue string) context.Context {
+	logger := log.LoggerFromContext(ctx)
+	logger = logger.WithField(fieldName, fieldValue)
+	return log.ContextWithLogger(ctx, logger)
+}
+
+func searchInSlice(length int, f func(i int) bool) (int, bool) {
+	for i := 0; i < length; i++ {
+		if f(i) {
+			return i, true
+		}
+	}
+	return -1, false
+}
